criando o V2
