--- lista.txt
@@ -1,4 +1,0 @@
-lab
-teste
-dashboard
-coisa
