+++ subdomain_finder.py
@@ -1,6 +1,6 @@
 """
 Usage:
-    subdomain_finder.py --target=<arg> --wordlist=<arg> [--threads=<arg>]
+    subdomain_finder.py --target=<arg> --wordlist=<arg> [--threads=<arg> --scan --uniq-ip]
     subdomain_finder.py --help
     subdomain_finder.py --version
 
@@ -14,128 +14,139 @@ Required options:
 
 Optional options:
     --threads=number of subdomais      how many subdomais test for threads [default: 100]
+    --scan                             scan subdomains (nmap)
+    --uniq-ip                          show list ips
 
 """
+
 import threading
 import socket
 
-from libnmap.process import NmapProcess
-from libnmap.parser import NmapParser, NmapParserException
+import nmap
 
 import os
 import sys
 from docopt import docopt, DocoptExit
 
 
+
 class myScan(threading.Thread):
-    
-     #TODO: Format output!
-    def __init__(self,target='localhost',param='',dict_domains={}):
+
+    def __init__(self, target='localhost', param='', dict_domains={}):
         threading.Thread.__init__(self)
         self.target = target
         self.param = param
         self.dict_domains = dict_domains
     
     def run(self):
-         report = self.scan_do(self.target, self.param)
-         if report:
-             self.scan_print(report)
-         else:
-             print("\n SCAN: No results returned for "+self.target)
-          
+        self.scan()
 
-    def scan_do(self,targets, options):
-         parsed = None
-         nmproc = NmapProcess(targets, options)
-         rc = nmproc.run()
-         if rc != 0:
-             print("\n SCAN: failed, {0}".format(nmproc.stderr))
-         #print(type(nmproc.stdout))
-         try:
-             parsed = NmapParser.parse(nmproc.stdout)
-         except NmapParserException as e:
-             print("\n SCAN: Exception raised while parsing scan: {0}".format(e.msg))
-     
-         return parsed
-     
-     
-    def scan_print(self,nmap_report):
-         
-         for host in nmap_report.hosts:
-             if len(host.hostnames):
-                 tmp_host = host.hostnames.pop()
-             else:
-                 tmp_host = host.address
+    def scan(self):
+        try:
+           nm = nmap.PortScanner()         # instantiate nmap.PortScanner object
+        except nmap.PortScannerError:
+            print('SCAN: Nmap not found', sys.exc_info()[0])
+            sys.exit(0)
+        except:
+            print("SCAN: Unexpected error:", sys.exc_info()[0])
+            sys.exit(0)
+        
+        scan_dict=nm.scan(self.target, self.param, arguments='-Pn -A')
+        print("##############################")
+        print("REPORT SCAN:")
+        print("    IP: "+self.target)
+        print("    OTHER SUB DOMAINS:")
+        for domain, ip in self.dict_domains.items():
+            if ip == self.target:
+                print("           "+domain)
 
-             print("##########################################")
-             print("SCAN REPORT FOR {0} ({1})".format(tmp_host,host.address))
-             print("    OTHER SUB DOMAINS:")
-             for domain, ip in self.dict_domains.items():
-                  if ip == host.address :
-                      print("           "+domain)
-             
-             #print("    STATUS HOST: {0}.".format(host.status))
-             print("    PORTS:")
-             print("          PORT     STATE         SERVICE")
-     
-             for serv in host.services:
-                 pserv = "        {0:>5s}/{1:3s}  {2:12s}  {3}".format(
-                         str(serv.port),
-                         serv.protocol,
-                         serv.state,
-                         serv.service)
-                 if len(serv.banner):
-                     pserv += " ({0})".format(serv.banner)
-                 print(pserv)
-     
-             if host.os_fingerprinted:
-                 print("\n    OS FINGERPRINT:")
-                 msg = ''
-                 for osm in host.os.osmatches:
-                     print("        Found Match:{0} ({1}%)".format(osm.name, osm.accuracy))
-                     #for osc in osm.osclasses:
-                     #    print("\tOS Class: {0}".format(osc.description))
-                     #    for cpe in osc.cpelist:
-                     #        print("\tCPE: {0}".format(cpe.cpestring))
-             else:
-                 print("\n    OS FINGERPRINT: No fingerprint available")
-             print("##########################################")
-
+        try:
+                for osmatch in nm[self.target]['osmatch']:
+                    print('     OS:{0} - {1}%'.format(osmatch['name'], osmatch['accuracy']))
+                    print('           OsClass.type : {0}'.format(osmatch['osclass'][0]['type']))
+                    print('           OsClass.vendor : {0}'.format(osmatch['osclass'][0]['vendor']))
+                    print('           OsClass.osfamily : {0}'.format(osmatch['osclass'][0]['osfamily']))
+                    print('           OsClass.osgen : {0}'.format(osmatch['osclass'][0]['osgen']))
+                    #print('           OsClass.accuracy : {0}%'.format(osmatch['osclass'][0]['accuracy']))
+                     
+        except:
+            pass
 
 
 
 
 class myTestDomain (threading.Thread):
-    def __init__(self,bDomain):
+    def __init__(self, bDomain):
         threading.Thread.__init__(self)
         self.bDomain = bDomain
-        self.dict_return={}
+        self.dict_return = {}
     
     def run(self):
         for domain in self.bDomain:
-            temp=self.finder(domain)
+            temp = self.finder(domain)
             if temp is not None:
-                self.dict_return[domain]=temp
+                self.dict_return[domain] = temp
     
-    def join( self ):
-        threading.Thread.join( self )
+    def join(self):
+        threading.Thread.join(self)
         return self.dict_return
 
     def finder(self, target):
             try:
                 ip = socket.gethostbyname(target)
                 if ip: 
-                        print("{0:65s} - {1}".format(target,ip))
+                        print("{0:65s} - {1}".format(target, ip))
                         return ip
             except:
                         pass
-    
 
 
 
+def subdomain_finder(threads, wordlist, target):
+
+    # Converting wordlist file to list
+    wlist = [line.rstrip('\n')+'.'+target for line in open(wordlist)]
+
+    len_chuck = int(threads)
+    # chunks of wordlist
+    chunks = [wlist[x:x+len_chuck] for x in range(0, len(wlist), len_chuck)]
+
+    # start threads
+    threads_domain = []
+    for chunk in chunks:
+        t=myTestDomain(chunk)
+        t.start()
+        threads_domain.append(t)
+
+    dict_domain = {}
+    for tw in threads_domain:
+        if len(tw.join()) > 0:
+            for domain, ip in tw.join().items():
+                dict_domain[domain] = ip
+
+    return dict_domain
 
 
+def ip_scan(domains_ips):
 
+    print('\n  SCAN... wait!')
+    threads_scan = []
+    for ip in ip_uniq(domains_ips):
+        scan = myScan(ip, "0-65535", domains_ips)
+        scan.start()
+        threads_scan.append(scan)
+
+    for ts in threads_scan:
+        ts.join()
+
+
+def ip_uniq(domains_ips):
+
+    ips = {}
+    for a, e in domains_ips.items():
+        ips[e] = 1
+
+    return ips
 
 def banner():
         os.system('clear')
@@ -153,60 +164,32 @@ def banner():
 
 def main():
     try:
+        banner()
         arguments = docopt(__doc__, version="TWA Corp. SubDomain Finder - 2016")
         target = arguments['--target']
         wordlist = arguments['--wordlist']
         threads = arguments['--threads']
+        opt_scan = arguments['--scan']
+        opt_uniq_ips = arguments['--uniq-ip']
 
     except DocoptExit as e:
         banner()
         os.system('python3 subdomain_finder.py --help')
         sys.exit(1)
 
+    domains_ips={}
+    domains_ips = subdomain_finder(threads, wordlist, target)
 
-    # Converting wordlist file to list
-    wlist = [line.rstrip('\n')+'.'+target for line in open(wordlist)]
-    
-    len_chuck = int(threads)
-    # chunks of wordlist
-    chunks = [wlist[x:x+len_chuck] for x in range(0, len(wlist), len_chuck)]
+    if opt_uniq_ips:
+        print("\n  IPs:")
+        ips = ip_uniq(domains_ips)
+        print("    Uniq: "+str(len(ips)))
+        for ip in ip_uniq(domains_ips):
+            print("        "+ip)
 
-
-    # start threads
-    threads_domain = []
-    for chunk in chunks:
-        t=myTestDomain(chunk)
-        t.start()
-        threads_domain.append(t)
-    
-    dict_domain={}
-    #waiting 
-    for tw in threads_domain:
-        if len(tw.join()) >0 :
-            for domain,ip in tw.join().items():
-                dict_domain[domain]=ip
-                
-
-    print ("\n  IPs:")
-    ips = {}
-    for a,e in dict_domain.items():
-        ips[e] = 1
-
-    print ("    Uniq: "+str(len(ips)))
-    for ip in ips.keys():
-        print("        "+ip)
-    
-
-    print('\n  SCAN... wait!')
-    threads_scan = []
-    for ip in ips.keys():
-        scan=myScan(ip,"-sV -A -Pn ",dict_domain)
-        scan.start()
-        threads_scan.append(scan)
-
-    for ts in threads_scan:
-        ts.join()
+    if opt_scan:
+        ip_scan(domains_ips)
 
 if __name__ == '__main__':
     main()
-#END...
+
